requirements.txt
pandas
numpy
scikit-learn
streamlit
tldextract
joblib
beautifulsoup4
requests
python-whois
matplotlib
seaborn
onnxruntime
huggingface_hub
playwright
fastapi
uvicorn[standard]
aiofiles
python-multipart
pydantic
tokenizers
